+++ MarkdownToFieldsFrontEditor.module.php
@@ -14,7 +14,7 @@ class MarkdownToFieldsFrontEditor extends WireData implements Module, Configurab
         return [
             'title' => 'MarkdownToFieldsFrontEditor',
             'summary' => 'Frontend editor for MarkdownToFields.',
-            'version' =>  '0.5.3',
+            'version' =>  '0.5.4',
             'autoload' => true,
             'singular' => true,
             'requires' => ['MarkdownToFields'],